# swagger.yaml
openapi: 3.0.0
info:
  title: Jobs API
  contact: {}
  version: '1.0'
servers:
- url: https://jobs-api-arie.herokuapp.com/api/v1
  variables: {}
paths:
  /jobs:
    get:
      tags:
      - Jobs
      summary: Get All Jobs
      operationId: GetAllJobs
      parameters:
      - name: address
        in: query
        description: address by exact province, region and city
        style: form
        explode: true
        schema:
          type: string
          example: 62447a2201456334f82a4af4
      - name: search
        in: query
        description: Fuzzy search based on title and description
        style: form
        explode: true
        schema:
          type: string
      - name: salaryRange
        in: query
        description: Determines the minimum salary and maximum salary
        style: form
        explode: true
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
...
components:
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
- httpBearer: []
...